Trac #33208: remove dead code in sage/rings/padics/pow_computer_ext.pyx.
The polynomial() method in this file declares a pointer "tmp", and
then reads the value of *tmp into the void before proceeding with the
actual body of the method that makes no use of "tmp" at all. This
raises a Cython warning,

  warning: sage/rings/padics/pow_computer_ext.pyx:797:8: local
  variable 'tmp' referenced before assignment

Here we delete those two purely ornamental lines.