Trac #8784: clean up after symmetrica when sage exits.
Symmetrica has an initialization ("anfang") routine that must be
called before using the library, and a teardown routine ("ende") that
you're supposed to call when you're done with it. The REPL calls
ende() via quit_sage() before it terminates, but users who run sage
scripts or use sage as a library will bypass that routine.

Here we remove ende() from quit_sage(), and instead register an atexit
hook for it. This way ende() is always called prior to exiting,
regardless of how sage was invoked. This is a step towards making
quit_sage() obsolete, and eventually removing it from sage.all.