Trac #8784: free the flint stack when sage terminates.
The quit_sage() function currently calls _fmpz_cleanup_mpz_content()
in an attempt to clean up some flint memory before the sage REPL
terminates. This is largely pointless because that function does
nothing in the thread-safe version of flint that most(?) distributions
provide. However, it's harmless, and a nice aid for tools like
valgrind that pester you about memory hygiene.

Here we move the call out of quit_sage() and into an atexit hook to be
run whenever sage quits. The main benefit of this is that the hook
will always be run, even when executing a *.sage script or utilizing
the sage library through python code.

This is a step towards making quit_sage() obsolete and eventually
removing it from sage.all.

This commit also removes the unused _fmpz_cleanup() function that
serves a similar purpose.