Trac #8784: free the integer pool when sage exits.
The sage.rings.integer module uses a global pool of integer objects
for performance reasons. The (empty) pool gets creates when the module
is imported, and is currently freed only when quit_sage() is called.
Freeing memory before termination is not strictly necessary, but is
a good practice and can aid tools like valgrind.

In this commit we move the cleanup operation out of quit_sage(), which
must be called manually, and into an atexit hook. This will ensure
that the pool is freed even when running a *.sage script or using the
sage library from python code.

This is a step towards making quit_sage() obsolete, and eventually
removing it from sage.all.