Trac #8784: clear quaternion algebra globals when sage exits.
Upon importing sage.algebras.quatalg.quaternion_algebra_element, the
_init_globals() function is called to initialize some variables. There
is a corresponding _clear_globals() that is presently only called
during quit_sage(). In this commit, we remove it from quit_sage() and
replace it by an atexit hook that is registered from within
_init_globals().

This ensures that _clear_globals() is always run, even when executing
a *.sage script or using the sage library from python code. This is a
step towards making quit_sage() obsolete, and eventually removing it
from sage.all.