Trac #8784: quit pexpect processes when sage terminates.
The quit_sage() function is currently responsible for terminating any
running pexpect processes when sage terminates. However, quit_sage()
is only invoked when exiting the REPL; not when processing a *.sage
script or while using the sage library from python code.

This commit moves the corresponding cleanup operation out of
quit_sage(), and into an atexit hook that is registered in the
_start() method of the Expect class. This should ensure that no rogue
processes remain after processing a sage script, as in Trac 33027, and
is a step towards making quit_sage() obsolete, and eventually removing
it from sage.all.

Note that we have replaced a single call to expect_quitall() with
several calls to the gentler quit(). This should still generally be
safe: the underlying close() method of pexpect's spawn class still
defaults to "force=True" and will send a SIGKILL if the process is
stubborn. Thus the "kill -9" attempted by kill_spawned_jobs() within
expect_quitall() is overkill.