Trac #8784: deprecate quit_sage().
When running *.sage scripts or using sage as a python library, the
user no longer needs to call quit_sage() to clean up after sage. All
of the cleanup tasks that used to live in quit_sage() have been moved
to atexit hooks that are invoked automatically. Thus quit_sage() is
obsolete and can be deleted from user code.

For now we mention that in a deprecation warning. Eventually this will
allow us to remove the top-level quit_sage() REPL command that,
confusingly, does not quit sage.