Trac #8784: cdef two internal quaternion functions.
The _init_globals() and _clear_globals() functions used by the
sage.algebras.quatalg.quaternion_algebra_element module are internal,
and should not be called manually: really, sage will segfault. Here we
change them from python functions to cdef (cython) functions so that
they are not visible outside of the module that uses them.

NB: I was a bit uneasy about passing a C function to python's
atexit.register(), but it seems to work.