Trac #33100: use a more reliable abs() replacement in is_hermitian().
Some finite fields for which is_hermitian() otherwise works are
missing abs(). Here we use the square root of z*z.conjugate() as a
replacement, which appears to work in all of the cases we care about.
Note that this would probably do the wrong thing if you tried to use
a nonzero tolerance with an (exact) finite field.