Trac #33293: don't run slow output replacements unconditionally.
Running regex substitutions on large outputs can slow our output
checker to a crawl. One of these in particular tries to fix some macOS
noise with a pattern involving several ".*" and is taking quite a
while to search through the string representation of a degree-99998
polynomial.

There is already a solution for this in sage.doctest.parsing called
_repr_fixups that pairs each fixup with a quick check to determine
whether or not it is necessary. This commit simply makes use of that
mechanism.