Trac #33293: add back some information to a polynomial_zmod_flint doctest.
A recent commit reworked a doctest output only the highest- and
lowest-degree nonzero terms of a polynomial, but those exponents
(99998 and 1, respectively) were hard-coded. As a result, the doctest
lost the information that those numbers are the degree and valuation
of the polynomial. Here we add that information back (thanks to Samuel
Lelièvre).