Trac #32981: cputime for pexpect interfaces on Linux/BSD.
After switching the doctests to report cputime instead of walltime,
the number for tests using pexpect interfaces are off because the time
spent in the subprocess is not accounted for. Fixing this turns out
not to be easy: the sage.misc.misc.cputime function can try to account
for them, but it is essentially asking the subprocess to compute its
own cputime. This has faults:

  1. Not every pexpect interface implemented (or can implement)
     such a method.

  2. The pexpect interface needs to be in good working condition
     before we can ask the subprocess to do anything. That's often
     not the case at the end of a "sage: ..." line in a doctest.
     Particularly it's not true in all of the tests for low-level
     details of our pexpect interface itself.

So, we compromise and parse /proc/<pid>/stat to obtain these
numbers. This works well the major caveat that the information is only
available on Linux and BSD. This is not *too* terrible: the motivation
for measuring these times in the doctest framework is that _someone_
will notice an outrageously long test and fix or report it. Moreover
the information is only missing for pexpect tests, and only "harmful"
for those tests that happen to run long.

The same information may be available on macOS through another
platform-specific interface.