Trac #33793: replace SAGE_TMP in SPYX_TMP.
With the move to a real temporary directory, we now also install an
atexit hook to clean up the tempdir that is created. Moreover
SPYX_TMP() is renamed to a new a @cached_function spyx_tmp() so that
only one directory will be created, and only if it's used.

The old SPYX_TMP path is marked with a deprecation warning.