Cythonize complex_plotting loops
This completes the implementation of colormaps and contour plotting of
complex plots. This commit cythonizes the lightness-adjusting maps. Now
the slow loops and function calls can all occur in C instead of python.

Please note that the testing code didn't change in this commit, and the
previous (just removed) code is easier to understand. I do this only for
speed.