Trac #24536: support complex expressions in find_{root,local_minimum}.
In concert with the previous commit, and similar to how this problem
was solved for plotting, we add support for symbolic complex
sub-expressions in find_root() and find_local_minimum() by using
complex fast-callables and by wrapping those fast-callables in the new
FastCallableFloatWrapper class.

The wrapper ignores imaginary noise below a configurable threshold
called "imaginary_tolerance" to return a real result. The
intermediate results however are allowed to remain complex.