Trac #24536: add backwards-compatibility case for "complex nan."
We have at least one function, find_root(), that expects to be able to
evaluate its objective function at the endpoints of an interval and
get back NaN if the function is undefined there. With a domain=float
fast-callable, that's what happens. But with domain=CDF, one gets
"NaN + Nan*I", and the FastCallableFloatWrapper doesn't know what to
do with that.

To keep everything working as it was, this commit adds a special case
for that situation, returning simply NaN in place of "NaN + Nan*I".