Trac #33363: shrink a hermite_form() test.
We have a test in sage.matrix.matrix_integer_dense that computes the
hermite_form() of a random 100-by-100 matrix. This usually runs
quickly, but sometimes takes a few minutes and can -- on a bad day --
take even longer: it once took 25m for me, which is how I noticed the
problem.

In this commit we shrink the size of the problem down to 15-by-15. At
that size, across thousands of random examples, the test always
completes quickly.