numpy: prefer venv headers before system headers
In #33473 it is explained how the headers for a system numpy will override the
headers for the numpy installed in the venv. When the system numpy and the venv
numpy are not binary compatible, this causes runtime errors.

The fix consists in always placing the venv numpy headers before the python
headers. This is done for `sagelib` and for `scipy.spatial` which are the two
modules that we know to be broken in this way.