Trac #32981: report both cpu and wall times for long tests.
We need to report the CPU time for tests that run afoul of
--warn-long, because --warn-long is now specified in CPU time. But
lacking an implementation of cputime that accounts for subprocesses on
macOS, we are still reporting the wall time in several places (such as
the per-file summary). So to keep things as consistent as possible, we
report the wall time for --warn-long failures as well.