doctest/control.py: add default for die_timeout
This was missing in the previous commit.

Also, tune up the timeout in the corresponding test in doctest/test.py.
This prevents a doctest failure when pytest is installed (since then
running `sage -t` will take ~ 3s after the doctest was actually run.

If #33531 is merged, the way we run `sage -t` here (using
`--optional=sage`) avoids the pytest overhead and the timeout can be
tuned down again.