Update to caching strategy in gauss_legendre
Numerics in Bruin-DisneyHogg-Gao suggest that the caching strategy previously implemented
for ``integrate_vector_N``, which is used in the Riemann surfaces module and currently
nowhere else, was not helpful so it is removed. This is done by renaming ``nodes`` to
``nodes_uncached``, retaining a cached version ``nodes`` for ``integrate_vector``, which
profits from the caching procedure.