#34714 fast path for factoring monomials over number fields
Trivial instances of univariate polynomial factorization over
complicated number fields are extremely slow because Sage starts by
trying to compute an integral basis.

We introduce a fast path that handles monomials — most importantly, the
identity polynomial.

Ideally, we should do something similar over arbitrary rings, but this
is currently difficult to implement in a generic because the factor()
method of constants is under-specified. In particular, calling factor()
on a rational number factors the numerator and denominator, which we
certainly do not want when factoring the leading coefficient of an
element of ℚ[x].

In the number field case, a possible generalization would be to detect
constant multiples of polynomials that factor over ℚ. Not sure if this
is a good idea.