Trac 33907: fix singular interface
This essentially reverts 85f65bf and a10d19d from trac #31846.
It turns out this was originaly written for #30945, but that issue was
fixed by upgrading cysignals.

Singular really needs a custom `_send_interrupt()` method, because the
default one will quit singular. Moreover, this handles two quirks of
singular:

 - a small delay before sending `chr(3)` works around a bug in singular.
 - sometimes one needs to send `;` a few times after interrupt to get
   back a prompt.

The original author of the custom `_send_interrupt()` is Jeroen Demeyer
in commit 17d23e9 (trac #10476). I changed the timeout for a smaller
one, and rewrote the doctest to call `interrupt()` explicitly instead of
using `alarm()` which introduces more noise.